Allow using tiny vae in other precisions
The model is available in fp16 only though

+++ nodes_model_loading.py
@@ -1649,9 +1649,9 @@ class WanVideoTinyVAELoader:
         model_path = folder_paths.get_full_path("vae_approx", model_name)
         vae_sd = load_torch_file(model_path, safe_load=True)
         
-        vae = TAEHV(vae_sd, parallel=parallel)
-       
-        vae.to(device = offload_device, dtype = dtype)
+        vae = TAEHV(vae_sd, parallel=parallel, dtype=dtype)
+
+        vae.to(device=offload_device, dtype=dtype)
 
         return (vae,)
